Docker update for rocBLAS, HCC, and HIP to 2.1 (#1606)
* Updated the requirements with newer HIP, HCC, rocBLAS, and clang-ocl

* Fixed boost issue.

* Fixed tidy issues across the lib.

* Formatting.

* Update hip and rocblas version.

* Fixed remaining tidy issues from formatting.

* Formatting.

* More tidy issues.

* Float conversion.

* NOLINT on stubborn coversion.

* ktime to float

* Remove unneeded NOLINT

* Removed unneeded semi-colon

* Drive updates.

* Switched to rocBLAS 2.3 RC commit hash.

* Use size_t for fusion op_map.

+++ src/batch_norm.cpp
@@ -81,7 +81,7 @@ TensorDescriptor BuildReshaped4DTensorDescriptor(const miopen::TensorDescriptor&
 void profileSequence(Handle& handle, unsigned char select, float* ctime)
 {
 
-    double ktime = 0.;
+    float ktime = 0.;
     assert((select < 3) && "profileSequence case incorrect");
     switch(select)
     {
