+++ src/ocl/mloNeuron.cpp
@@ -56,12 +56,12 @@ void mlo_construct_neuron::mloConstruct()
 {
     _hw_wave_sz = 64;
 
-    const int read_unit = 4;
+    const size_t read_unit = 4;
 
     size_t map_size = _search_params.in_width * _search_params.in_height * _search_params.n_inputs *
                       _search_params.batch_sz;
     size_t map_size_aligned = (map_size + read_unit - 1) / read_unit;
-    int N_PIXS_OFF          = map_size - (map_size / read_unit) * read_unit;
+    size_t N_PIXS_OFF       = map_size - (map_size / read_unit) * read_unit;
 
     size_t glbl_wk = map_size_aligned;
